Move common functions and types to separate file
